Remove z correction & add backend option
Older versions of the solver were adjusting the z position of a vertex after solving back
to lie on and not in between two voxel slices. This leads to issues in the core wise
processing scheme as the database is incosistent with the positions of vertices after
solving. This fixes the issue with branching violations.