Fix a bug that led to single vertices/edges not being forced if selected.
Random occurence of vertices that had 3 incoming edges were caused by the fact that in the conversion
of g1 to g2 graph we only considered full g2 vertices (3 vertices, 2 edges) to be forced. However, it
can also happen that there are single vertices or single edges present. These were ignored so far
leading to errors. This commit implements detection of these cases and forces the selection of at least
one of the g2 vertices that contain the forced edge or vertex.