Fix bug in interpolation
Removed the reduction to unique points at this cannot be done for each
edge individually. If there are positions besides start/end of each edge
that are duplicate this leads to problems as the graph used to be initilized
with the number of unique points. This is fixed by explicitly
calculating the number of start/end edge points that are repeated and
not relying on np.unique anymore.